+++ plone/tiles/data.py
@@ -1,35 +1,44 @@
 # -*- coding: utf-8 -*-
-
 import logging
 import urllib
-
-from zope.interface import implements, implementer, Interface
-from zope.component import adapts, adapter, queryUtility, getMultiAdapter
+from zope.interface import implements
+from zope.interface import implementer
+from zope.interface import Interface
+from zope.component import adapts
+from zope.component import adapter
+from zope.component import queryUtility
+from zope.component import getMultiAdapter
 from zope.component.interfaces import ComponentLookupError
-
-from zope.schema import getFieldsInOrder, getFields
+from zope.schema import getFieldsInOrder
+from zope.schema import getFields
 from zope.schema.interfaces import ISequence
-
 from zope.annotation.interfaces import IAnnotations
-
 from plone.tiles.interfaces import ITileType
 from plone.tiles.interfaces import ITile
 from plone.tiles.interfaces import IPersistentTile
 from plone.tiles.interfaces import ITileDataManager
 from plone.tiles.interfaces import ITileDataContext
 from plone.tiles.interfaces import IFieldTypeConverter
-
 from persistent.dict import PersistentDict
 
 try:
     import json
     assert json  # silence pyflakes
-except:
+except ImportError:
     import simplejson as json
 
 
 ANNOTATIONS_KEY_PREFIX = u'plone.tiles.data'
 LOGGER = logging.getLogger('plone.tiles')
+
+
+@adapter(ITile)
+@implementer(ITileDataManager)
+def transientTileDataManagerFactory(tile):
+    if tile.request.get('X-Tile-Persistent'):
+        return PersistentTileDataManager(tile)
+    else:
+        return TransientTileDataManager(tile)
 
 
 class TransientTileDataManager(object):
@@ -103,8 +112,23 @@ class PersistentTileDataManager(object):
 
         self.key = "%s.%s" % (ANNOTATIONS_KEY_PREFIX, tile.id,)
 
+    def _get_default_request_data(self):
+        # If we don't have a schema, just take the request
+        if self.tileType is None or self.tileType.schema is None:
+            data = self.tile.request.form.copy()
+        else:
+            # Try to decode the form data properly if we can
+            try:
+                data = decode(self.tile.request.form,
+                              self.tileType.schema, missing=True)
+            except (ValueError, UnicodeDecodeError,):
+                LOGGER.exception(u"Could not convert form data to schema")
+                return self.data.copy()
+        return data
+
     def get(self):
-        data = dict(self.annotations.get(self.key, {}))
+        data = self._get_default_request_data()
+        data.update(dict(self.annotations.get(self.key, {})))
         if self.tileType is not None and self.tileType.schema is not None:
             for name, field in getFields(self.tileType.schema).items():
                 if name not in data:
@@ -274,8 +298,8 @@ def decode(data, schema, missing=True):
 
             for item in value:
                 if isinstance(item, str):
-                    value = unicode(value, 'utf-8')
-                if not isinstance(item, field.value_type._type):
+                    value = unicode(item, 'utf-8')
+                if field.value_type._type and not isinstance(item, field.value_type._type):
                     item = value_type_field_type(item)
                 converted.append(item)